Fixes policy matching (#10)
Fixes policy matching

When matching policies against input resources, only the fields defined
in the policy should be considered. The input resource may contain a lot
of information describing itself, but policies may be interested only in
matching certain fields, and this needs to be taken into account when
matching policies.