Initial implementation of the Enforcer APIs
Adds an initial implementation of the following Enforcer APIs:

* Request Answer API
* Plugin API
* Controller API

It also provides a basic Plugin interface as long as a default
plugin loader implementation using the Go plugin mechanism. Plugin state
persistence is still to be implemented.

For policy storage, a local filesystem repository is implemented. The
current enforcer implementation can use it to store the current policy
bundle and to evaluate policies.

The Packet API is still to be implemented.